requirements.txt
numpy~=1.21.2
torch==1.10.0+cu111
torchaudio==0.10.0+cu111
matplotlib~=3.4.3
tensorboard~=2.6.0
librosa~=0.8.1
ipython~=7.29.0

pandas~=1.3.3
Pillow~=8.1.2
torchvision~=0.10.1
tqdm~=4.62.3
googledrivedownloader
gitpython